open ydr url for unlogged users, added tolltips for buy/sell buttons
