+++ src/components/TokenPanel/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import nextId from 'react-id-generator';
 import { observer } from 'mobx-react-lite';
 
@@ -21,9 +21,22 @@ interface IPanelContent {
 
 const TokenPanel: React.FC<TokenPanelProps> = observer(
   ({ panelContent, handleBuy, handleGetIn, handleSell }) => {
+    const [showTooltip, setShowTooltip] = useState(false);
+    const [pickTooltip, setPickTooltip] = useState<number | undefined>(undefined);
     // const walletConnector = useWalletConnectorContext();
     const { user } = useMst();
     const isTokenPicked = !!user.token;
+    const isUserLogged = !!localStorage?.yd_address || false;
+    const onClickHandler = (e: any) => {
+      e.preventDefault();
+      setShowTooltip(true);
+    };
+    const onFocusHandler = (value: number) => {
+      setPickTooltip(value);
+    };
+    const onBlurHandler = () => {
+      setShowTooltip(false);
+    };
     return (
       <div className="token-panel">
         <div className="token-panel__content">
@@ -37,12 +50,46 @@ const TokenPanel: React.FC<TokenPanelProps> = observer(
 
         <div className="token-panel__btns">
           {((isTokenPicked && handleBuy) || (!isTokenPicked && handleBuy && !handleGetIn)) && (
-            <Button className="token-panel__btn" onClick={handleBuy}>
+            <Button
+              className="token-panel__btn"
+              onClick={isUserLogged ? handleBuy : (e) => onClickHandler(e)}
+              onBlur={onBlurHandler}
+              onFocus={() => onFocusHandler(1)}
+            >
+              {pickTooltip === 1 && !isUserLogged && showTooltip && (
+                <div className="ant-tooltip ant-tooltip-placement-top">
+                  <div className="ant-tooltip-content">
+                    <div className="ant-tooltip-arrow">
+                      <span className="ant-tooltip-arrow-content" />
+                    </div>
+                    <div className="ant-tooltip-inner" role="tooltip">
+                      Please login
+                    </div>
+                  </div>
+                </div>
+              )}
               Buy
             </Button>
           )}
           {((isTokenPicked && handleSell) || (!isTokenPicked && handleSell && !handleGetIn)) && (
-            <Button className="token-panel__btn" onClick={handleSell}>
+            <Button
+              className="token-panel__btn"
+              onClick={isUserLogged ? handleSell : (e) => onClickHandler(e)}
+              onBlur={onBlurHandler}
+              onFocus={() => onFocusHandler(2)}
+            >
+              {pickTooltip === 2 && !isUserLogged && showTooltip && (
+                <div className="ant-tooltip ant-tooltip-placement-top">
+                  <div className="ant-tooltip-content">
+                    <div className="ant-tooltip-arrow">
+                      <span className="ant-tooltip-arrow-content" />
+                    </div>
+                    <div className="ant-tooltip-inner" role="tooltip">
+                      Please login
+                    </div>
+                  </div>
+                </div>
+              )}
               Sell
             </Button>
           )}
